Moved explore helpers to utils/explore
